requirements.txt
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0